move Serialize/Deserialize to CommonComparator

+++ PLC_esp8266/main/LogicProgram/MapIO.h
@@ -7,10 +7,10 @@ enum MapIO : uint8_t { DI = 0, AI, O1, O2, V1, V2, V3, V4 };
 extern const char *MapIONames[];
 
 static inline bool ValidateMapIO(MapIO mapIO) {
-    if (mapIO < MapIO::DI) {
+    if ((int8_t)mapIO < (int8_t)MapIO::DI) {
         return false;
     }
-    if (mapIO > MapIO::V4) {
+    if ((int8_t)mapIO > (int8_t)MapIO::V4) {
         return false;
     }
     return true;
